Convert RetrofitResponseTypeKeepProcessorTest to parametermized tests
