+++ retrofit-response-type-keeper/src/test/resources/all-http-methods/Service.pro
@@ -1,0 +1,10 @@
+# test.Service
+-keep,allowoptimization,allowshrinking,allowobfuscation class retrofit2.Call
+-keep,allowoptimization,allowshrinking,allowobfuscation class test.DeleteUser
+-keep,allowoptimization,allowshrinking,allowobfuscation class test.GetUser
+-keep,allowoptimization,allowshrinking,allowobfuscation class test.HeadUser
+-keep,allowoptimization,allowshrinking,allowobfuscation class test.HttpUser
+-keep,allowoptimization,allowshrinking,allowobfuscation class test.OptionsUser
+-keep,allowoptimization,allowshrinking,allowobfuscation class test.PatchUser
+-keep,allowoptimization,allowshrinking,allowobfuscation class test.PostUser
+-keep,allowoptimization,allowshrinking,allowobfuscation class test.PutUser
