+++ retrofit-response-type-keeper/src/test/resources/kotlin-suspend/Service.java
@@ -1,0 +1,11 @@
+package test;
+
+import kotlin.coroutines.Continuation;
+import retrofit2.*;
+import retrofit2.http.*;
+
+class Body {}
+
+interface Service {
+  @GET("/") Object get(Continuation<? extends Body> c);
+}
