+++ retrofit-response-type-keeper/src/test/resources/kotlin-suspend/Service.pro
@@ -1,0 +1,3 @@
+# test.Service
+-keep,allowoptimization,allowshrinking,allowobfuscation class java.lang.Object
+-keep,allowoptimization,allowshrinking,allowobfuscation class test.Body
