+++ retrofit-response-type-keeper/src/test/resources/nesting/Service.java
@@ -1,0 +1,12 @@
+package test;
+
+import retrofit2.*;
+import retrofit2.http.*;
+
+class One<T> {}
+class Two<T> {}
+class Three {}
+
+interface Service {
+  @GET("/") Call<One<Two<Three>>> get();
+}
