+++ retrofit-response-type-keeper/src/test/resources/nesting/Service.pro
@@ -1,0 +1,5 @@
+# test.Service
+-keep,allowoptimization,allowshrinking,allowobfuscation class retrofit2.Call
+-keep,allowoptimization,allowshrinking,allowobfuscation class test.One
+-keep,allowoptimization,allowshrinking,allowobfuscation class test.Three
+-keep,allowoptimization,allowshrinking,allowobfuscation class test.Two
